Rex module refactor* Renamed Rex to RexData* Extracted two new contexts from Entity
