--- .idea/.gitignore
@@ -1,9 +1,0 @@
-# Default ignored files
-/workspace.xml
-# Datasource local storage ignored files
-/dataSources/
-/dataSources.local.xml
-/vcs.xml
-/libraries
-dataSources.xml
-$PRODUCT_WORKSPACE_FILE$
