--- .idea/misc.xml
@@ -1,6 +1,0 @@
-<?xml version="1.0" encoding="UTF-8"?>
-<project version="4">
-  <component name="ProjectRootManager" version="2" languageLevel="JDK_10" default="false" project-jdk-name="Elixir 1.9.1" project-jdk-type="Elixir SDK">
-    <output url="file://$PROJECT_DIR$/classes" />
-  </component>
-</project>
