--- apps/rex/README.md
@@ -1,3 +1,0 @@
-# Rex
-
-**TODO: Add description**
