--- apps/rex/lib/rex.ex
@@ -1,9 +1,0 @@
-defmodule Rex do
-  @moduledoc """
-  Rex keeps the contexts that define your domain
-  and business logic.
-
-  Contexts are also responsible for managing your data, regardless
-  if it comes from the database, an external API or others.
-  """
-end
