--- apps/rex/lib/rex/repo.ex
@@ -1,5 +1,0 @@
-defmodule Rex.Repo do
-  use Ecto.Repo,
-    otp_app: :rex,
-    adapter: Ecto.Adapters.Postgres
-end
