--- apps/rex/rex.iml
@@ -1,57 +1,0 @@
-<?xml version="1.0" encoding="UTF-8"?>
-<module type="ELIXIR_MODULE" version="4">
-  <component name="NewModuleRootManager" inherit-compiler-output="true">
-    <content url="file://$USER_HOME$/Lab/Code/rex_umbrella/apps/rex">
-      <sourceFolder url="file://$USER_HOME$/Lab/Code/rex_umbrella/apps/rex/lib" isTestSource="false" />
-      <sourceFolder url="file://$USER_HOME$/Lab/Code/rex_umbrella/apps/rex/spec" isTestSource="true" />
-      <sourceFolder url="file://$USER_HOME$/Lab/Code/rex_umbrella/apps/rex/test" isTestSource="true" />
-      <excludeFolder url="file://$USER_HOME$/Lab/Code/rex_umbrella/apps/rex/assets/node_modules/phoenix" />
-      <excludeFolder url="file://$USER_HOME$/Lab/Code/rex_umbrella/apps/rex/assets/node_modules/phoenix_html" />
-      <excludeFolder url="file://$USER_HOME$/Lab/Code/rex_umbrella/apps/rex/cover" />
-      <excludeFolder url="file://$USER_HOME$/Lab/Code/rex_umbrella/apps/rex/deps" />
-      <excludeFolder url="file://$USER_HOME$/Lab/Code/rex_umbrella/apps/rex/doc" />
-      <excludeFolder url="file://$USER_HOME$/Lab/Code/rex_umbrella/apps/rex/logs" />
-    </content>
-    <orderEntry type="jdk" jdkName="Elixir 1.9.1" jdkType="Elixir SDK" />
-    <orderEntry type="sourceFolder" forTests="false" />
-    <orderEntry type="library" name="ecto_sql" level="project" />
-    <orderEntry type="library" name="postgrex" level="project" />
-    <orderEntry type="library" name="phoenix" level="project" />
-    <orderEntry type="library" name="ecto_enum" level="project" />
-    <orderEntry type="library" name="jason" level="project" />
-    <orderEntry type="library" name="telemetry" level="project" />
-    <orderEntry type="library" name="db_connection" level="project" />
-    <orderEntry type="library" name="ex_doc" level="project" />
-    <orderEntry type="library" name="benchee" level="project" />
-    <orderEntry type="library" name="benchee_json" level="project" />
-    <orderEntry type="library" name="decimal" level="project" />
-    <orderEntry type="library" name="connection" level="project" />
-    <orderEntry type="library" name="plug_cowboy" level="project" />
-    <orderEntry type="library" name="plug" level="project" />
-    <orderEntry type="library" name="phoenix_pubsub" level="project" />
-    <orderEntry type="library" name="inch_ex" level="project" />
-    <orderEntry type="library" name="gettext" level="project" />
-    <orderEntry type="library" name="phoenix_html" level="project" />
-    <orderEntry type="library" name="websocket_client" level="project" />
-    <orderEntry type="library" name="ecto" level="project" />
-    <orderEntry type="library" name="mariaex" level="project" />
-    <orderEntry type="library" name="earmark" level="project" />
-    <orderEntry type="library" name="cowboy" level="project" />
-    <orderEntry type="library" name="hackney" level="project" />
-    <orderEntry type="library" name="kadabra" level="project" />
-    <orderEntry type="library" name="x509" level="project" />
-    <orderEntry type="library" name="mime" level="project" />
-    <orderEntry type="library" name="plug_crypto" level="project" />
-    <orderEntry type="library" name="dialyze" level="project" />
-    <orderEntry type="library" name="benchee" level="project" />
-    <orderEntry type="library" name="benchee_json" level="project" />
-    <orderEntry type="library" name="phoenix_html" level="project" />
-    <orderEntry type="library" name="websocket_client" level="project" />
-    <orderEntry type="library" name="mariaex" level="project" />
-    <orderEntry type="library" name="earmark" level="project" />
-    <orderEntry type="library" name="hackney" level="project" />
-    <orderEntry type="library" name="kadabra" level="project" />
-    <orderEntry type="library" name="x509" level="project" />
-    <orderEntry type="library" name="dialyze" level="project" />
-  </component>
-</module>
