--- apps/rex/test/test_helper.exs
@@ -1,2 +1,0 @@
-ExUnit.start()
-Ecto.Adapters.SQL.Sandbox.mode(Rex.Repo, :manual)
