+++ apps/rex_data/.formatter.exs
