+++ apps/rex_data/lib/rex_data.ex
@@ -1,0 +1,9 @@
+defmodule RexData do
+  @moduledoc """
+  RexData module keeps the contexts that define your domain
+  and business logic.
+
+  Contexts are also responsible for managing your data, regardless
+  if it comes from the database, an external API or others.
+  """
+end
