+++ apps/rex_data/lib/rex_data/project/task.ex
@@ -1,0 +1,25 @@
+defmodule RexData.Project.Task do
+  use Ecto.Schema
+
+  import Ecto.Changeset
+
+  @derive {
+    Jason.Encoder,
+    only: [:frame, :project]
+  }
+  schema "tasks" do
+    field :frame, :integer
+    field :path, :string
+    field :project, :id
+    field :node, Ecto.UUID
+
+    timestamps()
+  end
+
+  @doc false
+  def changeset(task, attrs) do
+    task
+    |> cast(attrs, [:frame, :path, :project, :node])
+    |> validate_required([:frame, :project])
+  end
+end
