+++ apps/rex_data/lib/rex_data/repo.ex
@@ -1,0 +1,5 @@
+defmodule RexData.Repo do
+  use Ecto.Repo,
+    otp_app: :rex_data,
+    adapter: Ecto.Adapters.Postgres
+end
