+++ apps/rex_data/mix.exs
@@ -1,9 +1,9 @@
-defmodule RexWeb.MixProject do
+defmodule RexData.MixProject do
   use Mix.Project
 
   def project do
     [
-      app: :rex_web,
+      app: :rex_data,
       version: "0.1.0",
       build_path: "../../_build",
       config_path: "../../config/config.exs",
@@ -11,7 +11,6 @@ defmodule RexWeb.MixProject do
       lockfile: "../../mix.lock",
       elixir: "~> 1.5",
       elixirc_paths: elixirc_paths(Mix.env()),
-      compilers: [:phoenix, :gettext] ++ Mix.compilers(),
       start_permanent: Mix.env() == :prod,
       aliases: aliases(),
       deps: deps(),
@@ -24,7 +23,7 @@ defmodule RexWeb.MixProject do
   # Type `mix help compile.app` for more information.
   def application do
     [
-      mod: {RexWeb.Application, []},
+      mod: {RexData.Application, []},
       extra_applications: [:logger_file_backend, :logger, :runtime_tools]
     ]
   end
@@ -38,22 +37,26 @@ defmodule RexWeb.MixProject do
   # Type `mix help deps` for examples and options.
   defp deps do
     [
+      {:ecto_sql, "~> 3.1"},
+      {:postgrex, ">= 0.0.0"},
       {:phoenix, "~> 1.4.10"},
-      {:phoenix_pubsub, "~> 1.1"},
-      {:phoenix_ecto, "~> 4.0"},
-      {:gettext, "~> 0.11"},
-      {:rex_data, in_umbrella: true},
-      {:jason, "~> 1.1.2"},
-      {:plug_cowboy, "~> 2.0"},
-      {:logger_file_backend, "~> 0.0.10"}
+      {:ecto_enum, "~> 1.3"},
+      {:jason, "~> 1.0"},
+      {:logger_file_backend, "~> 0.0.11"}
     ]
   end
 
   # Aliases are shortcuts or tasks specific to the current project.
-  # For example, we extend the test task to create and migrate the database.
+  # For example, to create, migrate and run the seeds file at once:
+  #
+  #     $ mix ecto.setup
   #
   # See the documentation for `Mix` for more info on aliases.
   defp aliases do
-    [test: ["ecto.create --quiet", "ecto.migrate", "test"]]
+    [
+      "ecto.setup": ["ecto.create", "ecto.migrate", "run priv/repo/seeds.exs"],
+      "ecto.reset": ["ecto.drop", "ecto.setup"],
+      test: ["ecto.create --quiet", "ecto.migrate", "test"]
+    ]
   end
 end
