+++ apps/rex_web/lib/rex_web/handlers/node_handler.ex
@@ -7,13 +7,14 @@ defmodule RexWeb.NodeHandler do
   *handle_join*
   This method is used to handle node joining the rendering.
   """
-  alias Rex.{Entity, Manager, Utils}
+  alias RexData.{Node, Utils}
+  alias RexData.Node.NodeInfo
 
-  @spec handle_join(Node.t()) :: {:ok, map} | {:error, map}
+  @spec handle_join(NodeInfo.t()) :: {:ok, map} | {:error, map}
   def handle_join(%{"node_id" => node_id}) when node_id != "" do
     Logger.info("Trying to join with id: #{node_id}")
 
-    case Entity.get_node(node_id) do
+    case Node.get_node(node_id) do
       nil ->
         {:error,
          %{
@@ -31,7 +32,7 @@ defmodule RexWeb.NodeHandler do
     Logger.info("Generated UUID: #{node_id}")
 
     Map.put(payload, "node_id", node_id)
-    |> Entity.create_node()
+    |> Node.create_node()
     |> case do
       {:ok, _} ->
         Logger.info("Node #{node_id} has joined")
