--- apps/rex_web/rex_web.iml
@@ -1,41 +1,0 @@
-<?xml version="1.0" encoding="UTF-8"?>
-<module type="ELIXIR_MODULE" version="4">
-  <component name="NewModuleRootManager" inherit-compiler-output="true">
-    <content url="file://$USER_HOME$/Lab/Code/rex_umbrella/apps/rex_web">
-      <sourceFolder url="file://$USER_HOME$/Lab/Code/rex_umbrella/apps/rex_web/lib" isTestSource="false" />
-      <sourceFolder url="file://$USER_HOME$/Lab/Code/rex_umbrella/apps/rex_web/spec" isTestSource="true" />
-      <sourceFolder url="file://$USER_HOME$/Lab/Code/rex_umbrella/apps/rex_web/test" isTestSource="true" />
-      <excludeFolder url="file://$USER_HOME$/Lab/Code/rex_umbrella/apps/rex_web/assets/node_modules/phoenix" />
-      <excludeFolder url="file://$USER_HOME$/Lab/Code/rex_umbrella/apps/rex_web/assets/node_modules/phoenix_html" />
-      <excludeFolder url="file://$USER_HOME$/Lab/Code/rex_umbrella/apps/rex_web/cover" />
-      <excludeFolder url="file://$USER_HOME$/Lab/Code/rex_umbrella/apps/rex_web/deps" />
-      <excludeFolder url="file://$USER_HOME$/Lab/Code/rex_umbrella/apps/rex_web/doc" />
-      <excludeFolder url="file://$USER_HOME$/Lab/Code/rex_umbrella/apps/rex_web/logs" />
-    </content>
-    <orderEntry type="jdk" jdkName="Elixir 1.9.1" jdkType="Elixir SDK" />
-    <orderEntry type="sourceFolder" forTests="false" />
-    <orderEntry type="library" name="phoenix" level="project" />
-    <orderEntry type="library" name="phoenix_pubsub" level="project" />
-    <orderEntry type="library" name="phoenix_ecto" level="project" />
-    <orderEntry type="library" name="gettext" level="project" />
-    <orderEntry type="module" module-name="rex" />
-    <orderEntry type="library" name="jason" level="project" />
-    <orderEntry type="library" name="plug_cowboy" level="project" />
-    <orderEntry type="library" name="plug" level="project" />
-    <orderEntry type="library" name="telemetry" level="project" />
-    <orderEntry type="library" name="ex_doc" level="project" />
-    <orderEntry type="library" name="inch_ex" level="project" />
-    <orderEntry type="library" name="phoenix_html" level="project" />
-    <orderEntry type="library" name="websocket_client" level="project" />
-    <orderEntry type="library" name="dialyze" level="project" />
-    <orderEntry type="library" name="ecto" level="project" />
-    <orderEntry type="library" name="cowboy" level="project" />
-    <orderEntry type="library" name="hackney" level="project" />
-    <orderEntry type="library" name="kadabra" level="project" />
-    <orderEntry type="library" name="x509" level="project" />
-    <orderEntry type="library" name="mime" level="project" />
-    <orderEntry type="library" name="plug_crypto" level="project" />
-    <orderEntry type="library" name="decimal" level="project" />
-    <orderEntry type="library" name="earmark" level="project" />
-  </component>
-</module>
