--- scheme.json
@@ -1,29 +1,0 @@
-{
-  "id": "asdfsfwegweg",
-  "type": "animation",
-  "width": 1920,
-  "height": 1080,
-  "startingFrame": 1,
-  "frames": [
-    {
-      "id": 1,
-      "complexity": 1.20
-    },
-    {
-      "id": 2,
-      "complexity": 1.20
-    },
-    {
-      "id": 3,
-      "complexity": 1.20
-    },
-    {
-      "id": 4,
-      "complexity": 1.20
-    },
-    {
-      "id": 5,
-      "complexity": 1.20
-    }
-  ]
-}
